Move common to the src dir
