+++ src/apps/companies/controllers/__test__/archive.test.js
@@ -16,7 +16,7 @@ describe('Company controller, archive', () => {
         archiveCompany: this.stub.archiveCompany,
         unarchiveCompany: this.stub.unarchiveCompany,
       },
-      '../../../../config/logger': {
+      '../../../config/logger': {
         error: this.errorLoggerSpy,
       },
     })
