+++ src/apps/companies/transformers/__test__/company-to-about-view.test.js
@@ -2,7 +2,7 @@ const { aboutLabels } = require('~/src/apps/companies/labels')
 const { NOT_SET_TEXT } = require('~/src/apps/companies/constants')
 
 const transformCompanyToAboutView = proxyquire('~/src/apps/companies/transformers/company-to-about-view', {
-  '../../../../config': {
+  '../../../config': {
     currencyRate: {
       usdToGbp: 0.75,
     },
