+++ src/apps/events/__test__/transformers.test.js
@@ -10,7 +10,7 @@ const {
   transformEventResponseToViewRecord,
   transformEventFormBodyToApiRequest,
 } = proxyquire('~/src/apps/events/transformers', {
-  '../../../config': {
+  '../../config': {
     archivedDocumentsBaseUrl,
   },
 })
