+++ src/common/transform-sectors.js
