+++ src/lib/__test__/date.test.js
@@ -1,4 +1,4 @@
-const { parseDateString } = require(`~/common/date`)
+const { parseDateString } = require(`~/src/common/date`)
 
 describe('Date tests', () => {
   describe('parse human dates', () => {
