+++ src/middleware/__test__/features.test.js
@@ -17,7 +17,7 @@ describe('feature flag middleware', () => {
     this.loggerStub = sinon.stub()
 
     this.features = proxyquire('~/src/middleware/features', {
-      '../../config/logger': {
+      '../config/logger': {
         error: this.loggerStub,
       },
     })
